requirements.txt
websocket-client
pillow
python-multipart
fastapi
uvicorn
websocket-client
termcolor
requests
python-dotenv